Modified unit tests
