+++ MangaManager/CoverManagerLib/.pytest_cache/v/cache/lastfailed
@@ -1,3 +1,4 @@
 {
+  "tests.py": true,
   "tests.py::CbzControllerTester": true
 }
